support dynamic library on gcc
A few notes: boost was not built with fPIC. Dynamic library cannot
reference it statically in that case, probably unless we make some
leaky dependency mechanism where consumers implicitly link to
dependencies of dynamic libraries as well.

Also notice the visibility("default") in the IMPORT macro. This
surprised me, but in my testing, it seems as though compiling the
consumer with visibility("hidden") causes the linker to not find
the symbol in the shared library where it's defined.