Change naming scheme for version/build type
Naming debug/release binaries the same doesn't really work if both are
going to be installed in the same location. It's easier to distinguish
if there are different names. Now, debug binaries have ".debug" in them.

Moreover, executables should have a major version in the name, similar
to "python3", etc.  If a user doesn't like having to type the version
number, a symlink or alias is a good option to link a versionless name.
However, if a script relies on a specific version, this allows two
versions to be included in the same PATH and have predictable behavior.

Only the major version is included in the name. The most up to date
minor version should be the one that is installed on the system. It's
the job of the installer/package manager to handle these upgrades.