start recursively including dynamic dependencies
With a heavy heart, dynamic library dependencies are NOT private if
a template function in a library uses a compiled function in a
dependent library.  Windows already has leaky dynamic libraries, so
it seems silly to not have the default that dynamic library dependencies
are leaky. A new feature can fully support private library dependencies
if/when necessary.