(v0.1.1) Some major issues were fixed
- source maps were added for better debugging
- some basic build tracing is available via GULPACHEK_DEBUG_LOG env var
- track paths being built instead of target instances

It's possible that we should only "untrack" a dependency after the top
level build is done. This seems a bit idealistic for now. We currently
track while a target is built while it's being built. If we have a
problem where a target is being built too many times due to starting
after that in-progress instance was done, that means it isn't updating
the mtime and it's rebuilding anyways, and that's what should ultimately
be addressed.