Add feature for marking params for target
This is an additional way for the build system to figure out that a
dependency needs to be built.  A target can override params() and return
an object that parametrizes the recipe, independent of the dependencies
that the target already lists.  More technically, a target will now can
have a dependency on a file whose contents are a hash of the object
returned by params(), and this dependency will follow the normal rules
of dependencies.