Auto generate dependency for which path-based dependencies are attached
It became clear in an early use case that the actual dependencies are
often parameters to the recipe, so it seemed worth supporting this
automatically: if a build path's target's dependencies change, the
target probably needs updating.

We're only using the target's dependencies with paths currently. If the
user uses info from dependencies for the recipe that changes in path
name (and regular mtime content update) would not catch, add these
explicitly to params()