Improve path generation api / concept. Pattern for build function
The new way of writing a Recipe doesn't give the Recipe a way to
expose it's dependencies as being recipes as well. If all the
consumer does is add() to the cookbook, then the dependencies won't
also be added. This means either the Cookbook needs to ask something
else of the Recipe or someone needs to register with it via add().

Nothing is preventing build libraries from exposing higher level APIs
to add() to the cookbook, so this is deemed better because it keeps
the Recipe API simpler.  It seems nice from a consumer standpoint as
well to not have to `new` an object but function calls are more
concise.  It also leaves more of the implementation up to the library,
so more complicated registrations could happen. See the addSass()
file in example/make.js in this commit.