Simplify IRecipe type
The juice wasn't worth the squeeze for the automatic typings.

It's cleaner and simpler and more predictable to pass in the
generic type of the implementation to args.paths<MyType>() than
to specify the type both in the IRecipe<MyType> implementation and
in the IRecipeBuildArgs<MyType> and know that this is what drives
the paths to have strong typing.  This is simpler for everyone since
it's more concise and predictable to use, and it's easier to implement.