Rework completed recipes to show target groups
The behavior changed in that only targets whose recipe
was completed will now show.  It used to show for any
target that was evaluated as a dependency. Not sure which
one I like better yet. Visual and non-breaking change so
not a huge deal to get right in this 'release'.